6-8: agree with moving phrase start to third-last eighth note of 8 instead of overlapping with I.
I put a V tonicization here because however brief I felt that it was not a half cadence as the original V/V/V to V/V to V model suggests. The predominant and dominant preparation makes this a PAC in V in my opinion. Alternatively we could consider using I/V instead of V at the end.